// packages/plugins/paramfinder/src/types.ts
export type HeaderMap = Record<string, string[]>;

export interface EngineRequest {
  id: string;
  host: string;
  port: number;
  url: string;
  path: string;
  query: string;
  method: string;
  headers: HeaderMap;
  body: string;
  tls: boolean;
  raw: string;
  context: RequestContext;
}

export interface EngineResponse {
  status: number;
  headers: HeaderMap;
  body: string;
  time: number;
  length: number;
  raw?: string;
}

export interface EngineRequestResponse {
  request: EngineRequest;
  response: EngineResponse;
}

export interface Parameter {
  name: string;
  value: string;
}

export const AnomalyType = {
  StatusCode: "status-code",
  Headers: "headers",
  ReflectionCount: "reflection_count",
  Body: "body",
  Redirect: "redirect",
  Similarity: "similarity",
} as const;

export type AnomalyType = (typeof AnomalyType)[keyof typeof AnomalyType];

export interface StatusCodeAnomaly {
  type: typeof AnomalyType.StatusCode;
  from: number;
  to: number;
}

export interface RedirectAnomaly {
  type: typeof AnomalyType.Redirect;
  from?: string;
  to?: string;
}

export interface HeadersAnomaly {
  type: typeof AnomalyType.Headers;
  headerName: string;
  from?: string[];
  to?: string[];
}

export interface ReflectionCountAnomaly {
  type: typeof AnomalyType.ReflectionCount;
  parameterName: string;
  from: number;
  to: number;
}

export interface BodyLengthAnomaly {
  type: typeof AnomalyType.Body;
  check: "length";
  from: number;
  to: number;
}

export interface BodyContentAnomaly {
  type: typeof AnomalyType.Body;
  check: "content";
  expectedDiffCount: number;
  actualDiffCount: number;
}

export interface SimilarityAnomaly {
  type: typeof AnomalyType.Similarity;
  similarity: number;
  threshold: number;
}

export type Anomaly =
  | StatusCodeAnomaly
  | RedirectAnomaly
  | HeadersAnomaly
  | ReflectionCountAnomaly
  | BodyLengthAnomaly
  | BodyContentAnomaly
  | SimilarityAnomaly;